aider: Refactored obtaining `location_id` from `patient_data` to fetching it from the encounter data in `src/services/sync_service.py`.
# Aider chat conversation:

ASSISTANT: To address the issue with the missing `location_id` in the `patient_data` dictionary, we need to modify the `src/services/sync_service.py` file. We will need to find an alternative way to obtain the `location_id` for the patient's encounters and use it to map to the DHIS2 `orgUnit` ID.

Here is the *SEARCH/REPLACE block* for the change:

src/services/sync_service.py
```python
<<<<<<< SEARCH
                "orgUnit": location_mappings.get(str(patient_data['location_id'])),  # Map location ID to DHIS2 orgUnit ID
=======
                # Fetch the location ID from the encounter data instead of patient_data
                "orgUnit": location_mappings.get(str(self.openmrs_connector.get_location_id_by_encounter_id(encounter_id))),
>>>>>>> REPLACE
```

This change assumes that we have a method `get_location_id_by_encounter_id` in the `OpenMRSConnector` class that can fetch the location ID given an encounter ID. If such a method does not exist, it will need to be implemented in the `OpenMRSConnector` class.

Please confirm if you would like to proceed with this change, and if the `OpenMRSConnector` class has the method `get_location_id_by_encounter_id` implemented. If it is not implemented, I can provide the necessary code to add this method.