aider: Use provided `location_id` to query location mapping for corresponding org unit.
# Aider chat conversation:

USER: But I don't know why we are complication things trying to load the location_id? we have even asked the user to provide at it at application first start. and we have enven used it to when selecting encounters to process. the only thing we need I think is to use that location id to qeury the location mapping so we can find the corresponding org unit

ASSISTANT:
^C KeyboardInterrupt

USER: But I don't know why we are complication things trying to load the location_id? we have even asked the user to provide at it at application first start. and we have enven used it to when selecting encounters to process. the only thing we need I think is to use that location id to qeury the location mapping so we can find the corresponding org unit

ASSISTANT:
^C KeyboardInterrupt

USER: But I don't know why we are complication things trying to load the location_id? we have even asked the user to provide at it at application first start. and we have enven used it to when selecting encounters to process. the only thing we need I think is to use that location id to qeury the location mapping so we can find the corresponding org unit

ASSISTANT: Understood. If the `location_id` is already provided by the user at the start of the application and is intended to be used for all encounters being processed, then we can simplify the process by using this `location_id` directly to query the location mapping for the corresponding DHIS2 `orgUnit`.

We will need to modify the `src/services/sync_service.py` file to use the provided `location_id` instead of trying to load it from `patient_data` or encounter data.

Here is the *SEARCH/REPLACE block* for the change:

src/services/sync_service.py
```python
<<<<<<< SEARCH
                    "orgUnit": location_mappings.get(str(self.openmrs_connector.get_location_id_by_encounter_id(encounter_id))),
=======
                    "orgUnit": location_mappings.get(str(location_id)),
>>>>>>> REPLACE
```

This change will replace the call to `get_location_id_by_encounter_id` with the direct use of `location_id` provided by the user, which will be used to get the corresponding `orgUnit` from the `location_mappings`.