aider: Remove the `trackedEntityInstance` key and its value from the `patients_to_sync.json` file.
# Aider chat conversation:

ASSISTANT: To remove the `trackedEntityInstance` from the `patients_to_sync.json` file, we need to modify the file and remove the key-value pair corresponding to `trackedEntityInstance`. Here is the *SEARCH/REPLACE block* for this change:

patients_to_sync.json
```json
<<<<<<< SEARCH
{
    "trackedEntityInstance": "247620",
=======
{
>>>>>>> REPLACE
```

Please note that this change will remove the `trackedEntityInstance` key and its value from the JSON object. If there are multiple instances of `trackedEntityInstance` in the file and you want to remove all of them, please let me know, and I will adjust the instructions accordingly.