aider: The condition for setting the value of `Citizenship` or `country` to `D6Us3GQryHU` has been added in the `process_patient_and_encounters` method in `src/services/sync_service.py`.
# Aider chat conversation:

USER: In process_patient_and_encounters in `src/services/sync_service.py` if openmrs_attr equals to Citizenship or country, I want to set its' value to D6Us3GQryHU . Can you add that condition just bellow where you added a check for Sex? And `src/services/sync_service.py` is already in the chat, don't create another.

ASSISTANT: To add the condition for setting the value of `Citizenship` or `country` to `D6Us3GQryHU` in the `process_patient_and_encounters` method in `src/services/sync_service.py`, we will modify the code as follows:

src/services/sync_service.py
```python
<<<<<<< SEARCH
# Special handling for gender to convert 'F' to 'Female' and 'M' to 'Male'
if openmrs_attr == 'Sex':
    patient_attribute_value = 'Female' if patient_attribute_value == 'F' else 'Male' if patient_attribute_value == 'M' else patient_attribute_value
# Map OpenMRS attributes to DHIS2 attributes and append them to the attributes list
if patient_attribute_value is not None:
    dhis2_compliant_json["attributes"].append({
        "attribute": dhis2_attr,
        "value": patient_attribute_value
    })
=======
# Special handling for gender to convert 'F' to 'Female' and 'M' to 'Male'
# and for 'Citizenship' or 'country' to set its value to 'D6Us3GQryHU'
if openmrs_attr == 'Sex':
    patient_attribute_value = 'Female' if patient_attribute_value == 'F' else 'Male' if patient_attribute_value == 'M' else patient_attribute_value
elif openmrs_attr in ['Citizenship', 'country']:
    patient_attribute_value = 'D6Us3GQryHU'
# Map OpenMRS attributes to DHIS2 attributes and append them to the attributes list
if patient_attribute_value is not None:
    dhis2_compliant_json["attributes"].append({
        "attribute": dhis2_attr,
        "value": patient_attribute_value
    })
>>>>>>> REPLACE
```

This change adds an `elif` condition to check if the `openmrs_attr` is either `Citizenship` or `country` and, if so, sets the `patient_attribute_value` to `D6Us3GQryHU`. The rest of the code remains unchanged, and the modified value is then appended to the `attributes` list in the `dhis2_compliant_json` object.